scanner: Move scanner.token to token.token
This commit creates a new token package and
moves scanner.token to it.  This is a stepping
stone towards adding the parser, which also
depends on the Token type.  Obviously Token
could have stayed in scanner, but now that it
has two dependencies moving it to its own package
seems structurally more sound.